requirements.txt
streamlit==1.31.0
fireworks-ai==0.8.1
python-dotenv==1.0.0
pydantic==2.4.2
beautifulsoup4==4.12.3
requests==2.31.0
scrapy==2.11.0
langchain==0.1.0
pandas==2.1.4
numpy==1.26.3
tqdm==4.66.1
python-dateutil==2.8.2
scikit-learn==1.4.0
